requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
httpx>=0.28,<1.0
qdrant-client>=1.15,<2.0
langgraph>=0.6,<1.0
docling>=2.53,<3.0
pypdf>=5.5,<6.0
python-pptx>=1.0,<2.0
Pillow>=11.0,<12.0
sympy>=1.14,<2.0
scipy>=1.16,<2.0
numpy>=2.3,<3.0
control>=0.10,<1.0
